% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_concentrated_poverty}
\alias{add_concentrated_poverty}
\title{Joins a Column for Concentrated Poverty}
\usage{
add_concentrated_poverty(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}
}
\value{
A tibble with column \code{concentrated_poverty} added to the data of
the input tibble
}
\description{
The \code{add_concentrated_poverty()} function obtains the protocol Concentrated
Poverty Index based on data extracted from the U.S. Census Bureau’s American
Community Survey (ACS) 5-year estimates. Concentrated poverty describes the
proportion of individuals below the poverty level living in \strong{high-poverty}
neighborhoods, typically defined as those with poverty rates above 40\%
(\strong{PhenX uses 35\%}). \href{https://data.census.gov/table/ACSST5Y2019.S1701?q=S1701}{Data.census.gov}.
}
\details{
Table S1701 - Poverty Status in the past 12 months
\itemize{
\item Calculation:
\item Total individuals below poverty = sum of \code{S1701_C02_046E} across all areas
\item Total individuals below poverty in high-poverty areas = sum of \code{S1701_C02_046E} in areas where \code{S1701_C03_046E} > 35\%
\item Total individuals below poverty in high-poverty areas / Total individuals below poverty
}
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_concentrated_poverty(test_tib)

}
\references{
Source:

U.S. Census Bureau. (2018). Poverty status in the past 12 months (table S1701), 2009-2018 American Community Survey 5-year estimates. \href{https://data.census.gov/cedsci/table?q=table\%20S1701&hidePreview=false&tid=ACSST5Y2018.S1701&vintage=2018}{Retrieved from}
U.S. Census Bureau. (2018). Ratio of income to poverty level in the past 12 months (table C17002), 2009-2018 American Community Survey 5-year estimates. \href{https://data.census.gov/cedsci/table?q=C17002&hidePreview=false&tid=ACSDT5Y2018.C17002&vintage=2018}{Retrieved from}

General Reference:

Jargowsky, P. A. (1997). Poverty and place: Ghettos, barrios, and the American city. Russell Sage Foundation.
Jargowsky, P. A. (2003). Stunning progress, hidden problems: The dramatic decline of concentrated poverty in the 1990s (Living Cities Census Series). Brookings Institution Center on Urban and Metropolitan Policy.
Jargowsky, P. A. (2013). Concentration of poverty in the new millennium: Changes in the prevalence, composition, and location of high-poverty neighborhoods. The Century Foundation and Rutgers Center for Urban Research and Education.
}
