% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_decennial_dissimilarity}
\alias{add_decennial_dissimilarity}
\title{Joins a Column for Residential Segregation: Decennial Dissimilarity}
\usage{
add_decennial_dissimilarity(
  tib,
  minority_group_code_dhc = NULL,
  minority_group_code_sf1 = NULL
)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}

\item{minority_group_code_dhc}{The P5 table code of the comparison group of
interest. If the code isn't known leave null and select from the menu. Used
when dates are 2020 and up. \href{https://www.census.gov/data/tables/2023/dec/2020-census-dhc.html}{DHC data}}

\item{minority_group_code_sf1}{The P005 table code of the comparison group of
interest. If the code isn't known leave null and select from the menu. Used
when dates are below 2020. \href{https://www.census.gov/data/datasets/2010/dec/summary-file-1.html}{sf1 data}}
}
\value{
A tibble with columns \code{decennial_year} and
\code{decennial_dissimilarity_index} added to the data of the input tibble
}
\description{
The function \code{add_decennial_dissimilarity()} computes the
\code{decennial_dissimilarity_index} (D) for a given dataset containing
race/ethnic population counts within subareas (\verb{Census tracts}) of a larger
area (\code{County}). The Dissimilarity Index is a commonly used measure of
residential segregation, reflecting the evenness of distribution of two
population groups across geographic areas. Accurate decennial year data
starts at 2010, so dates early are rounded and all other dates round normally
to nearest decennial year. For the decennial years of 2020, a column for
their 2010 data is also added. This is because not all of 2020 data is loaded
since the decennial year is not over, so to assure proper analysis 2010 can
be used to compare with.
}
\details{
The Dissimilarity Index (D) is calculated using the formula:

\deqn{D = \frac{1}{2} \sum_{i=1}^{n} \left| \frac{x_i}{X} - \frac{y_i}{Y} \right|}

where:
\itemize{
\item \eqn{x_i} is the minority population in subarea
\item \eqn{X} is the total minority population in the region
\item \eqn{y_i} is the reference population in subarea
\item \eqn{Y} is the total reference population in the region
\item \eqn{n} is the number of subareas (tracts) in the region
}

The index ranges from 0 (complete integration) to 1 (complete segregation).
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_decennial_dissimilarity(test_tib)

}
\references{
Source:

Recommended data sources include:

U.S. Census Bureau decennial Census (1990, 2000, and 2010), \link{available from} (https://www.census.gov/programs-surveys/decennial-census/data.html).
U.S. Census Bureau. (1991). 1990 Census of Population and Housing, Summary Tape File 1, Technical Documentation. \href{https://www2.census.gov/programs-surveys/decennial/1990/technical-documentation/complete-tech-docs/summary-files/d1-d90-s100-14-tech.zip}{Available from}.
U.S. Census Bureau. (2001). Census 2000, Summary File 1, Technical Documentation,\href{https://www2.census.gov/programs-surveys/decennial/2000/technical-documentation/complete-tech-docs/summary-files/sf1.pdf}{ Available from}.
U.S. Census Bureau. (2012). 2010 Census Summary File 1, Technical Documentation. \href{https://www2.census.gov/programs-surveys/decennial/2010/technical-documentation/complete-tech-docs/summary-file/sf1.pdf}{Available from}.
\href{https://data.census.gov}{Census website}.

Note that several online sources provide Dissimilarity Index scores for selected metropolitan statistical areas, counties, and school districts (and across Census years). See, for example, the American Communities Project and the School Segregation Project at the Brown and Lewis Mumford Center at Albany (https://www.brown.edu/academics/spatial-structures-in-social-sciences/).

General References:

Iceland, J., & Douzet, F. (2006). Measuring racial and ethnic segregation. Hrodote, 122(3): 25-43.
Iceland, J., Weinberg, D. H., & Steinmetz, E. (2002). Racial and ethnic residential segregation in the United States: 1980-2000 (U.S. Census Bureau, Series CENSR‑3). Washington DC: U.S. Government Printing Office.
Massey, D. S., & Denton, N. A. (1988). The dimensions of residential segregation. Social Forces,67, 281-315.
Morgan, P.M., Murphy, R.F., Willis, R.A., Hubbard, D.W., & Norton, J.M. (1975). Dental health of Louisiana residents based on the ten-state nutrition survey. Public Health Reports, 90(2), 173-178.
Morrill, R.L. (1995). Aging in place, age specific migration and natural decrease. The Annals of Regional Science, 29(1), 41-66.
Reardon, S. F. (2006). A conceptual framework for measuring segregation and its associations with population outcomes. In J. M. Oakes & J. S. Kaufman (Eds.), Methods in social epidemiology (pp. 169-192). San Francisco, CA: Wiley and Sons/Jossey-Bass.
Reardon, S. F., & Firebaugh, G. (2002). Measures of multi-group segregation. Sociological Methodology, 32, 33-67.
Reardon, S. F., Matthews, S. A., O’Sullivan, D., Lee, B. A., Firebaugh, G., Farrell, C. R., & Bischoff, K. (2008). The geographic scale of metropolitan racial segregation. Demography,45(3), 489-514.
Reardon, S. F., & O’Sullivan, D. (2004). Measures of spatial segregation. Sociological Methodology, 34, 121-162.
Sakoda, J.M. (1981). A generalized index of dissimilarity. Demography, 18(2), 245-50.
Taeuber, K. E., & Taeuber, A. F. (1965). Negroes in cities: Residential segregation and neighborhood change. Chicago, IL: Aldine.
Theil, H. (1972). Statistical decomposition analysis (Vol. 14). Amsterdam, The Netherlands: North-Holland.
White, M. J. (1983). The measurement of spatial segregation. American Journal of Sociology,88, 1008-1018.
White, M. J. (1986). Segregation and diversity measures in population distribution. Population Index,52, 198-221.
Wong, D. S. (1993). Spatial indices of segregation. Urban Studies, 30, 559-572.
}
