% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_dissimilarity_index}
\alias{add_dissimilarity_index}
\title{Joins a Column for Residential Segregation: Dissimilarity Index}
\usage{
add_dissimilarity_index(tib, minority_group_code = NULL)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}

\item{minority_group_code}{The B03002 table code of the minority group of
interest. If the code isn't known leave null and select from the menu}
}
\value{
A tibble with column \code{dissimilarity_index} added to the data of the
input tibble
}
\description{
The \code{add_dissimilarity_index()} function calculates the \code{dissimilarity_index}
to assess residential segregation between two population groups (typically a
\strong{minority group} and a \strong{reference group}, non-Hispanic whites in most
studies) using U.S. Census Bureau data. Residential segregation reflects the
distribution of different race/ethnic groups across \emph{census tracts} within a
larger area, \emph{County} in this case. The Dissimilarity Index is a commonly
used measure of segregation that quantifies how evenly two groups are
distributed across geographic areas.

This implementation uses ACS 5-year estimates for race/ethnic population
counts extracted from Table B03002: "Hispanic or Latino by Race." A higher
index value indicates greater segregation, ranging from 0 (complete
integration) to 1 (complete segregation).
\href{https://data.census.gov/table/ACSDT1Y2022.B03002}{Data.census.gov}.
}
\details{
The Dissimilarity Index (D) is calculated as:
\deqn{D = \frac{1}{2} \sum_{i=1}^n \left| \frac{x_i}{X} - \frac{y_i}{Y} \right|}
where:
\itemize{
\item \eqn{x_i} is the minority group population in tract i
\item \eqn{X} is the total minority population across all tracts
\item \eqn{y_i} is the reference group population in tract i
\item \eqn{Y} is the total reference group population across all tracts
}
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_dissimilarity_index(test_tib)

}
\references{
Source:

Recommended data sources include the following:

U.S. Census Bureau American Community Survey (ACS) data products (5-year estimates). \href{http://www.census.gov/programs-surveys/acs}{Source}
\href{http://factfinder.census.gov}{American Factfinder website}

Note that several online sources provide Dissimilarity Index scores for selected metropolitan statistical areas, counties, and school districts (and across Census years). See, for example, the \link{American Communities Project and the School Segregation Project at the Brown and Lewis Mumford Center at Albany}(http://www.s4.brown.edu/cen2000/index.html: http://www.s4.brown.edu/schoolsegregation/index.htm) as well as the \href{http://rtispatialdata.rti.org/}{Spatial Impact Factor Web Data at RTI International}.

General References:

Iceland, J., & Douzet, F. (2006). Measuring racial and ethnic segregation. Hrodote, 122(3): 25-43.
Iceland, J., Weinberg, D. H., & Steinmetz, E. (2002). Racial and ethnic residential segregation in the United States: 1980-2000 (U.S. Census Bureau, Series CENSR‑3). Washington DC: U.S. Government Printing Office.
Massey, D. S., & Denton, N. A. (1988). The dimensions of residential segregation. Social Forces, 67, 281-315.
Morgan, P. M., Murphy, R. F., Willis, R. A., Hubbard, D. W., & Norton, J. M. (1975). Dental health of Louisiana residents based on the ten-state nutrition survey. Public Health Reports, 90(2), 173-178.
Morrill, R. L. (1995). Aging in place, age specific migration and natural decrease. Annals of Regional Science, 29(1), 41-66.
Reardon, S. F. (2006). A conceptual framework for measuring segregation and its associations with population outcomes. In J. M. Oakes & J. S. Kaufman (Eds.), Methods in social epidemiology (pp. 169-192). San Francisco, CA: Wiley and Sons/Jossey-Bass.
Reardon, S. F., & Firebaugh, G. (2002). Measures of multi-group segregation. Sociological Methodology, 32, 33-67.
Reardon, S. F., Matthews, S. A., O’Sullivan, D., Lee, B. A., Firebaugh, G., Farrell, C. R., & Bischoff, K. (2008). The geographic scale of metropolitan racial segregation. Demography, 45(3), 489-514.
Reardon, S. F., & O’Sullivan, D. (2004). Measures of spatial segregation. Sociological Methodology, 34, 121-162.
Sakoda, J. M. (1981). A generalized index of dissimilarity. Demography, 18(2), 245-250.
Taeuber, K. E., & Taeuber, A. F. (1965). Negroes in cities: Residential segregation and neighborhood change. Chicago, IL: Aldine.
Theil, H. (1972). Statistical decomposition analysis (Vol. 14). Amsterdam, Netherlands: North-Holland.
White, M. J. (1983). The measurement of spatial segregation. American Journal of Sociology, 88, 1008-1018.
White, M. J. (1986). Segregation and diversity measures in population distribution. Population Index, 52, 198-221.
Wong, D. S. (1993). Spatial indices of segregation. Urban Studies, 30, 559-572.
}
