% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_education_attainment}
\alias{add_education_attainment}
\title{Joins a Column for Education Attainment}
\usage{
add_education_attainment(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}
}
\value{
A tibble with column \code{education_attainment} added to the data of the
input tibble
}
\description{
This function extracts and processes educational attainment data from the
U.S. Census Bureau’s American Community Survey (ACS) to assess the
educational environment of a respondent's geographic area. The measure is
based on the percentage of individuals with a bachelor’s degree or higher in
a given census tract. The data are sourced from ACS 5-year estimates and
updated annually. The function requires a geocoded address to determine the
respondent's location and linka it to the appropriate ACS statistics. Users
can access ACS data from the U.S. Census Bureau via \href{https://data.census.gov/table/ACSDT5Y2020.B15003}{Data.census.gov}.
}
\details{
Variables are extracted from ACS Table B15003 - Education Attainment
for population 25 years old +
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_education_attainment(test_tib)

}
\references{
Source:

U.S. Census Bureau. (n.d.). Data.census.gov. \href{http://data.census.gov}{Retrieved from}
U.S. Census Bureau. (n.d.). American Community Survey (ACS) products (specifically, the 5-year estimates). \href{http://www.census.gov/programs-surveys/acs}{Retrieved from}

General References:

Agency for Healthcare Research and Quality (AHRQ). (2015). Understanding the relationship between education and health: A review of the evidence and an examination of community perspectives. \href{http://www.ahrq.gov/professionals/education/curriculum-tools/population-health/zimmerman.html}{Retrieved from}
Arcaya, M., Tucker-Seeley, R., Kim, R., Schnake-Mahl, A., So, M., & Subramanian, S.V. (2016). Research on neighborhood effects on health in the United States: A systematic review of study characteristics. Social Science & Medicine, 168, 16–29.
Cohen, S. S., Mumma, M. T., Dupree, E. E., & Boice, J. D., Jr. (2018). Validating the use of census data on education as a measure of socioeconomic status in an occupational cohort. International Journal of Radiation Biology, 19, 1–10.
Connelly, R., Gayle, V., & Lambert, P. S. (2016). A review of educational attainment measures for social survey research. Methodological Innovations, 9.
Friedman, J., Graetz, N., & Gakidou, E. (2018). Improving the estimation of educational attainment: New methods for assessing average years of schooling from binned data. PLoS One, 13(11), e0208019.
Geronimus, A. T., & Bound, J. (1998). Use of census-based aggregate variables to proxy for socioeconomic group: Evidence from national samples. American Journal of Epidemiology, 148(5), 475–486.
Krueger, P. M., Dehry, I. A., & Chang, V. W. (2019). The economic value of education for longer lives and reduced disability. Milbank Quarterly, 97(1), 48–73.
McElroy, J. A., Remington, P. L., Trentham-Dietz, A., Robert, S. A., & Newcomb, P. A. (2003). Geocoding addresses from a large population-based study: Lessons learned. Epidemiology, 14, 399–407.
Pardo-Crespo, M. R., Narla, N. P., Williams, A. R., Beebe, T. J., Sloan, J., Yawn, B. P., … Juhn, Y. J. (2013). Comparison of individual-level versus area-level socioeconomic measures in assessing health outcomes of children in Olmsted County, Minnesota. Journal of Epidemiology and Community Health, 67(4), 305–310. PMID: 23322850
Ryan, C. L., & Bauman, K. (2016). Educational attainment in the United States: 2015. P20-578, 1–11. \href{https://www.census.gov/library/publications.html}{Retrieved from}
}
