% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_EJ_index}
\alias{add_EJ_index}
\title{Joins a Column for Environmental Justice Indices}
\usage{
add_EJ_index(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}
}
\value{
A tibble with columns \code{diesal_pm}, \code{particulate_matter}, \code{ozone},
\code{lead_paint}, \code{traffic_prox_vol}, \code{prox_to_NPLS}, \code{prox_to_RMPS},
\code{prox_to_TSDF}, \code{prox_to_MDWD}, \code{NATA_cancer}, and \code{NATA_resp_haz} added to
the data of the input tibble
}
\description{
⚠️ \strong{Note:} Data is retrieved from \href{https://zenodo.org/records/14767363}{archived data}.The US EPA EJSCREEN tool and
data was discontinued around May 2024 and remains unavailable. Some \code{GEOID}
values will not have matches in data, possibly due to discontinuation or
missing information for that year.

The \code{add_EJ_index()} function appends Environmental Justice (EJ) index data
from the U.S. EPA's EJSCREEN tool to a user-provided tibble, joining by
\code{GEOID} and \code{actual_year} columns. It selects and merges environmental
indicators based on whether \code{GEOID} exists and is available based on the
EJSCREEN year (2016–2024).

Environmental indicators added:
\itemize{
\item National Scale Air Toxics Assessment Diesel PM (diesal_pm)
\item Particulate Matter (PM2.5) (particulate_matter)
\item Ozone (ozone)
\item Lead Paint Indicator (lead_paint)
\item Traffic Proximity and Volume (traffic_prox_vol)
\item Proximity to National Priorities List Sites (prox_to_NPLS)
\item Proximity to Risk Management Plan Sites (prox_to_RMPS)
\item Proximity to Treatment Storage and Disposal Facilities (prox_to_TSDF)
\item Proximity to Major Direct Water Dischargers (prox_to_MDWD)
\item NATA Air Toxics Cancer Risk (NATA_cancer)
\item NATA Respiratory Hazard Index (NATA_resp_haz)
}

If a year is outside this range or missing, the function defaults to the closest available year.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_EJ_index(test_tib)

}
\references{
Source:

U.S. Environmental Protection Agency (EPA). (2018). EJSCREEN: Environmental Justice Screening and Mapping Tool. \href{https://www.epa.gov/ejscreen}{Retrieved from}.
Demographic information that is obtained from the U.S. Census Bureau’s American Community Survey (ACS). The 2018 version of EJSCREEN includes 2012–2016 ACS 5-year summary file data, which are based on 2014 census boundaries.

General References:

Cifuentes, P., Reichard, J., Im, W., Smith, S., Colen, C., Giurgescu, C., … Hood, D. B. (2019). Application of the Public Health Exposome Framework to estimate phenotypes of resilience in a model Ohio African-American women's cohort. Journal of Urban Health, 96(Suppl 1), 57–71.
Driver, A., Mehdizadeh, C., Bara-Garcia, S., Bodenreider, C., Lewis, J., & Wilson, S. (2019). Utilization of the Maryland Environmental Justice Screening Tool: A Bladensburg, Maryland case study. International Journal of Environmental Research and Public Health, 16(3).
Rowangould, D., Rowangould, G., Craft, E., & Niemeier, D. (2018). Validating and refining EPA's traffic exposure screening measure. International Journal of Environmental Research and Public Health, 16(1).
U.S. Environmental Protection Agency (EPA). (2019). Technical Documentation for EJSCREEN. \href{https://www.epa.gov/ejscreen/technical-documentation-ejscreen}{Retrieved from}.
}
