% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_ICE}
\alias{add_ICE}
\title{Joins a Column for Residential Concentrations of Income (ICE)}
\usage{
add_ICE(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}
}
\value{
A tibble with column \code{ICE} added to the data of the input tibble
}
\description{
The \code{add_ICE()} function computes the Index of Concentration(ICE) for census
tracts based on U.S. Census Bureau American Community Survey (ACS) 5-year
estimates. ICE measures the extent to which an area’s residents are
concentrated into groups at the extremes of deprivation and privilege. Values
range from \strong{-1} (\emph{100\% in most deprived}) to \strong{1} (\emph{100\% in most
privileged}). Data is collected from \href{https://data.census.gov/table/ACSDT5Y2021.B19001}{Table B19001}.
}
\details{
The ICE is calculated using ACS variables representing income
distribution within census tracts. Specifically, it uses income percentiles
to identify the proportion of affluent and poor residents in each tract and
computes ICE as \strong{(A_i - P_i) / T_i}, where \strong{A_i} is the number of
affluent persons,
\strong{P_i} is the number of poor persons, and \strong{T_i} is the total population with known income.

This metric can be used to assess socioeconomic polarization in neighborhoods
or larger areas, aiding research on social determinants of health and
neighborhood effects.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_ICE(test_tib)

}
\references{
Source:

U.S. Census Bureau ACS data products (5-year estimates). \href{https://data.census.gov/cedsci/}{source}.
Data.census.gov website. \href{https://data.census.gov/}{source}.
Massey, D. S. (2001). The prodigal paradigm returns: ecology comes back to sociology. In: A. Booth & A. Crouter (Eds.), Does it take a village? community effects on children, adolescents, and families (pp. 41-48). Lawrence Erlbaum Associates.

General References:

Chambers, B. D., Baer, R. J., McLemore, M. R., & Jelliffe-Pawlowski, L. L. (2019). Using Index of Concentration at the Extremes as indicators of structural racism to evaluate the association with preterm birth and infant mortality-California, 2011-2012. Journal of Urban Health, 96(2), 159-170. \href{https://doi.org/10.1007/s11524-018-0272-4}{source}.
Feldman, J. M., Waterman, P. D., Coull, B. A., & Krieger, N. (2015). Spatial social polarization: using the Index of Concentration at the Extremes jointly for income and race/ethnicity to analyse risk of hypertension. Journal of Epidemiology and Community Health, 69(12), 1199-1207. \href{https://doi.org/10.1136/jech-2015-205728}{source}.
Krieger, N., Waterman, P. D., Gryparis, A., & Coull, B. A. (2015). Black carbon exposure, socioeconomic and racial/ethnic spatial polarization, and the Index of Concentration at the Extremes (ICE). Health & Place, 34, 215-228. \href{https://doi.org/10.1016/j.healthplace.2015.05.008}{source}.
Krieger, N., Kim, R., Feldman, J., & Waterman, P. D. (2018). Using the Index of Concentration at the Extremes at multiple geographical levels to monitor health inequities in an era of growing spatial social polarization: Massachusetts, USA (2010-14). International Journal of Epidemiology, 47(3), 788–819. \href{https://doi.org/10.1093/ije/dyy004}{source}.
Krieger, N., Waterman, P. D., Spasojevic, J., Li, W., Maduro, G., & Van Wye, G. (2016). Public health monitoring of privilege and deprivation with the Index of Concentration at the Extremes. American Journal of Public Health, 106(2), 256-263. \href{https://doi.org/10.2105/AJPH.2015.302955}{source}.
}
