% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_min_wage}
\alias{add_min_wage}
\title{Joins a Column for Minimum Wage}
\usage{
add_min_wage(tib, current_fed_min_wage = 7.25)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}

\item{current_fed_min_wage}{The current years minimum wage (e.g. current year: 2025, $7.25)}
}
\value{
A tibble with columns \code{mw_year}, \code{above_federal}, \code{ever_above}, \code{dichot_inter_mw}, and \code{mw_above_fed} added to the data of the input tibble
}
\description{
The function \code{add_min_wage()}joins state minimum wage data to a dataset
containing state and year variables. It merges in whether the state's minimum
wage was \emph{above} the \strong{federal minimum wage} in a \emph{given year}, whether it \strong{ever} had
a \emph{higher} minimum wage, an \strong{interaction} of the two(multiply), and the \strong{difference} between
the \emph{state} and \emph{federal} minimum wage. These variables are constructed for use
in analyses of minimum wage policies.

*NOTE: The years are modified as \code{mw_year} to account for missing years in the records(if needed)
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

clean_min_wage_data(7.25)

add_min_wage(test_tib)

}
\references{
Source:

Merrill-Francis, M., Vernick, J. S., McGinty, E. E., & Pollack Porter, K. M. (2022). Association between fatal occupational injuries and state minimum-wage laws, 2003-2017. American Journal of Preventive Medicine, 62(6), 878-884. \href{https://doi.org/10.1016/j.amepre.2021.09.022}{source}.
Wage and Hour Division. (2022, January). State Minimum Wage Laws. U.S. Department of Labor, Wage and Hour Division. \href{https://www.dol.gov/agencies/whd/minimum-wage/state}{source}.

General References:

Avila, C. J., & Frakt, A. B. (2021). Raising the minimum wage and public health. JAMA Health Forum, 2(1), e201587. \href{https://doi.org/10.1001/jamahealthforum.2020.1587}{source}.
Buszkiewicz, J. H., Hill, H. D., & Otten, J. J. (2021). Association of state minimum wage rates and health in working-age adults using the National Health Interview Survey. American Journal of Epidemiology, 190(1), 21-30. \href{https://doi.org/10.1093/aje/kwaa018}{source}.
Kuehn, David. (2014). The importance of study design in the minimum-wage debate (Issue Brief #384). Economic Policy Institute. \href{https://www.epi.org/publication/importance-study-design-minimum-wage-debate/}{source}.
Leigh, J. Paul, & Du, Juan. (2018, October 4). Effects of minimum wages on population health. Health Policy Brief. Health Affairs. \href{https://doi.org/10.1377/hpb20180622.107025}{source}.
Neumark, D., & Wascher, W. (2006). Minimum wages and employment: A review of evidence from the new minimum wage research. Working Papers 060708, University of California-Irvine, Department of Economics, revised Jan 2007. \href{https://ideas.repec.org/p/irv/wpaper/060708.html}{source}.
Paul Leigh, J., Leigh, W. A., & Du, J. (2019). Minimum wages and public health: A literature review. Preventive Medicine, 118(122-134). \href{https://doi.org/10.1016/j.ypmed.2018.10.005}{source}.
}
