% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_pct_unionized}
\alias{add_pct_unionized}
\title{Joins a Column for Percent Unionized}
\usage{
add_pct_unionized(tib, current_year = 2025, default_year = 2024)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}

\item{current_year}{The current year}

\item{default_year}{The year that NA values should be set to if greater than
current year or NA by the given input tibble}
}
\value{
A tibble with columns \code{union_year}, \code{mem_unions}, and \code{rep_unions} added to the data of the input tibble
}
\description{
This function adds percent unionized data for non-agricultural workers
to a tibble, using state and year-based union affiliation data from
the \href{https://www.bls.gov/cps/cpslutabs.htm}{Current Population Survey (CPS)}.
}
\details{
The data is sourced from the Bureau of Labor Statistics, based on union
membership(\code{mem_unions}) and representation(\code{rep_unions}) estimates. Data includes the percentage of
employed wage and salary workers who are members of or represented by unions.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_pct_unionized(test_tib)

}
\references{
Source:

Bureau of Labor Statistics. (2007). Labor force statistics from the Current Population Survey. Retrieved May 22, 2019, from \href{https://www.bls.gov/cps/cpslutabs.htm}{source}

General References:

Halpern-Manners, A., & Warren, J. R. (2012). Panel conditioning in longitudinal studies: Evidence from labor force items in the Current Population Survey. Demography, 49(4), 1499–1519. doi: 10.1007/s13524-012-0124-x
Hirsch, B. T., & Macpherson, D. A. (n.d.). Union membership and earnings data book: Compilations from the Current Population Survey. Washington, DC: Bureau of National Affairs. Updated annually to the present.
Hirsch, B. T., & Macpherson, D. A. (2003). Union membership and coverage database from the Current Population Survey: Note. Industrial and Labor Relations Review, 56(2), 349–554. Updated annually at \href{www.unionstats.com}{source}.
Milkman, R., & Luce, S. (2017). Labor unions and the Great Recession. Russell Sage Foundation Journal of the Social Sciences, 3(3), 145–165.
Perusek, G. (2018). U.S. union membership data in perspective. Retrieved May 22, 2019, from \href{https://newlaborforum.cuny.edu/2018/03/03/u-s-union-membership-data-perspective/}{source}
Walker, A. N. (2014). Labor's enduring divide: The distinct path of public sector unions in the United States. Studies in American Political Development, 28(2), 175–200.
U.S. Census Bureau. (2018). Current Population Survey, 2018 Annual Social and Economic (ASEC) Supplement (machine-readable data file). Conducted for the Bureau of Labor Statistics, Washington, DC.
}
