% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_RFEI}
\alias{add_RFEI}
\title{Joins a Column for Food Swamp: Retail Food Environment Index}
\usage{
add_RFEI(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}
}
\value{
A tibble with column \code{RFEI} added to the data of the input tibble
}
\description{
The \code{add_RFEI()} function computes the Retail Food Environment Index (RFEI)
at the county level for each participant in a provided tibble, using store
and restaurant data from the \href{https://www.ers.usda.gov/data-products/food-environment-atlas/data-access-and-documentation-downloads}{USDA Food Environment Atlas}.
The index is calculated as the ratio of fast food and convenience stores to
grocery stores and supermarkets. Data is matched by year (\strong{2011} or
\strong{2016}), state, and cleaned county name. All values that will cause issues
due to NA are taken care of. The number indicates the times greater of
unhealthy food access. A higher RFEI has been correlated with a higher
obesity rate compared to a lower RFEI.
}
\details{
\deqn{RFEI = \frac{FFR + CONVS}{GROC + SUPERC}}
where:
\itemize{
\item \code{FFR} = Number of fast food restaurants (limited-service restaurants)
\item \code{CONVS} = Number of convenience stores
\item \code{GROC} = Number of grocery stores
\item \code{SUPERC} = Number of supermarkets
}
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_RFEI(test_tib)

}
\references{
Source:

Cooksey-Stowers, K., Schwartz, M. B., & Brownell, K. D. (2017). Food swamps predict obesity rates better than food deserts in the United States. International Journal of Environmental Research and Public Health, 14(11), 1366.

General References:

California Center for Public Health Advocacy, PolicyLink, & the UCLA Center for Health Policy Research. (2008, April). Designed for disease: The link between local food environments and obesity and diabetes. \href{https://escholarship.org/uc/item/9zc7p54b}{Retrieved from}.
Cooksey-Stowers, K., Schwartz, M. B., & Brownell, K. D. (2017). Food swamps predict obesity rates better than food deserts in the United States. International Journal of Environmental Research and Public Health, 14(11), 1366.
Economic Research Service (ERS), U.S. Department of Agriculture (USDA). (2019). Data access and documentation downloads. \href{https://www.ers.usda.gov/data-products/food-environment-atlas/data-access-and-documentation-downloads/}{Retrieved from}.
Economic Research Service (ERS), U.S. Department of Agriculture (USDA). (2019). Food Environment Atlas. \href{https://www.ers.usda.gov/data-products/food-environment-atlas/go-to-the-atlas/}{Retrieved from}
McGuirt, J. T., Jilcott Pitts, S. B., & Gustafson, A. (2018). Association between spatial access to food outlets, frequency of grocery shopping, and objectively-assessed and self-reported fruit and vegetable consumption. Nutrients, 10(12), 1974.
Murphy, M., Badland, H., Jordan, H., Koohsari, M. J., & Giles-Corti, B. (2018). Local food environments, suburban development, and BMI: A mixed methods study. International Journal of Environmental Research and Public Health, 15(7), 1392.
Lucan, S. C., Maroko, A. R., Seitchik, J. L., Yoon, D. H., Sperry, L. E. & Schechter, C. B. (2018). Unexpected neighborhood sources of food and drink: Implications for research and community health. American Journal of Preventive Medicine, 55(2), e29–e38.
Spence, J. C., Cutumisu, N., Edwards, J., Raine, K. D., & Smoyer-Tomic, K. (2009). Relation between local food environments and obesity among adults. BMC Public Health, 9, 192. doi: 10.1186/1471-2458-9-192
}
