% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_separation_index}
\alias{add_separation_index}
\title{Joins a Column for Residential Segregation: Separation Index}
\usage{
add_separation_index(tib, comparison_group_code = NULL)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}

\item{comparison_group_code}{The B03002 table code of the comparison group of
interest. If the code isn't known leave null and select from the menu}
}
\value{
A tibble with column \code{separation_index} added to the data of the
input tibble
}
\description{
The \code{add_separation_index()} function extracts race/ethnicity data from the
U.S. Census Bureau's American Community Survey (ACS) using the \code{tidycensus}
package and calculates Separation Index (S) for measuring residential
segregation at the \emph{census tract} level within a given \emph{County}.

The indices are calculated using Table B03002: "Hispanic or Latino by Race"
to reflect the evenness of race/ethnic group distribution across smaller
geographic units relative to a larger area. \href{https://data.census.gov/table/ACSDT1Y2022.B03002}{Data.census.gov}.
}
\details{
\deqn{p_i = \frac{n_{1i}}{n_{1i} + n_{2i}}}

where \eqn{n_{1i}} is the tract-level estimate of the reference group and
\eqn{n_{2i}} is that of the comparison group.

Similarly, the proportion of the reference group in the county is calculated
as:

\deqn{P = \frac{N_1}{N_1 + N_2}}

where \eqn{N_1} and \eqn{N_2} are the county-level totals of the reference
and comparison groups, respectively.

The separation index is then defined as the absolute difference between the
tract-level and county-level proportions:

\deqn{S_i = |p_i - P|} The separation index ranges from 0 to 1: 0 being no
separation and 1 being complete separation.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_separation_index(test_tib)

}
\references{
Source:

Brown University. (2019). Spatial structures in social sciences. Retrieved May 28, 2019, \href{https://www.brown.edu/academics/spatial-structures-in-social-sciences/}{from}
Fossett, M. (2017). New methods for measuring and analyzing segregation. Springer.
U.S. Census Bureau. (2019). American Community Survey (ACS) data products (5-year estimates). Retrieved May 28, 2019, \href{http://www.census.gov/programs-surveys/acs}{from}
U.S. Census Bureau. (2019). American FactFinder. Retrieved May 28, 2019, \href{http://factfinder.census.gov}{from}

General References:

Carrington, W. J., & Troske, K. R. (1997). On measuring segregation in samples with small units. Journal of Business & Economic Statistics, 15(4), 402–409.
Fossett, M. (2017). New methods for measuring and analyzing segregation. Springer.
Iceland, J., & Douzet, F. (2006). Measuring racial and ethnic segregation. Herodote, 122(3), 25–43.
Iceland, J., Weinberg, D. H., & Steinmetz, E. (2002). Racial and ethnic residential segregation in the United States: 1980-2000 (U.S. Census Bureau, Series CENSR-3). Washington, DC: U.S. Government Printing Office.
James, D. R., & Taeuber, K. E. (1985). Measures of segregation. Sociological Methodology, 15, 1–32.
Massey, D. S., & Denton, N. A. (1988). The dimensions of residential segregation. Social Forces, 67, 281–315.
Morgan, P. M., Murphy, R. F., Willis, R. A., Hubbard, D. W., & Norton, J. M. (1975). Dental health of Louisiana residents based on the ten-state nutrition survey. Public Health Reports, 90(2), 173–178.
Morrill, R. L. (1995). Aging in place, age specific migration and natural decrease. Annals of Regional Science, 29(1), 41–66.
Reardon, S. F. (2006). A conceptual framework for measuring segregation and its associations with population outcomes. In J. M. Oakes & J. S. Kaufman (Eds.), Methods in social epidemiology (pp. 169–192). San Francisco, CA: Wiley and Sons/Jossey-Bass.
Reardon, S. F., & Firebaugh, G. (2002). Measures of multi-group segregation. Sociological Methodology, 32, 33–67.
Reardon, S. F., Matthews, S. A., O’Sullivan, D., Lee, B. A., Firebaugh, G., Farrell, C. R., & Bischoff, K. (2008). The geographic scale of metropolitan racial segregation. Demography, 45(3), 489–514.
Reardon, S. F., & O’Sullivan, D. (2004). Measures of spatial segregation. Sociological Methodology, 34, 121–162.
Sakoda, J. M. (1981). A generalized index of dissimilarity. Demography, 18(2), 245–250.
Taeuber, K. E., & Taeuber, A. F. (1965). Negroes in cities: Residential segregation and neighborhood change. Chicago, IL: Aldine.
Theil, H. (1972). Statistical decomposition analysis (vol. 14). Amsterdam, Netherlands: North-Holland.
White, M. J. (1983). The measurement of spatial segregation. American Journal of Sociology, 88, 1008–1018.
White, M. J. (1986). Segregation and diversity measures in population distribution. Population Index, 52, 198–221.
Winship, C. (1977). A revaluation of indexes of residential segregation. Social Forces, 55(4), 1058–1066.
Wong, D. S. (1993). Spatial indices of segregation. Urban Studies, 30, 559–572.
Zoloth, B. S. (1976). Alternative measures of school segregation. Land Economics, 52(3), 278–298.
}
