% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{add_SVI}
\alias{add_SVI}
\title{Joins a Column for Social Vulnerability(SVI)}
\usage{
add_SVI(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}
}
\value{
A tibble with columns \code{SVI_rank_meth} and \code{SVI_flag_meth} added to
data of the input tibble
}
\description{
The \code{add_SVI()} function joins and run calculations with ACS data to a
user-supplied tibble using \code{GEOID} and \code{year} as matching keys. The SVI
measures the relative vulnerability of populations in U.S. Census tracts
based on 15 social factors grouped into 4 themes. There are two methods
ranking and flagging. Ranking gives a percentile rank compared to all other
tracts in that same state (0-1), and flagging gives the score of factors that
are greater than or equal to the 90th percentile for that state (0-15).
Flagging is broader and quicker, whereas ranking is more in-depth. Both are
provided with this function.
}
\details{
Tables used:
\itemize{
\item \href{https://data.census.gov/table/ACSST5Y2019.S1701?q=S1701}{S1701} ~ Population below poverty
\item \href{https://data.census.gov/table/ACSST1Y2021.S2301?g=040XX00US53_050XX00US53057_010XX00US&text=Table+s2301}{S2301} ~ Unemployment count
\item \href{https://data.census.gov/table/ACSDT1Y2021.B19301?q=per+capita+income}{B19301} ~ Per capita
\item \href{https://data.census.gov/table/ACSST1Y2022.S1501}{S1501} ~ No Highschool diploma(25+)
\item \href{https://data.census.gov/table/ACSST1Y2022.S0101}{S0101} ~ 65 years old +
\item \href{https://data.census.gov/table/ACSST1Y2022.S0101}{S0101} ~ Under 18 years old
\item \href{https://data.census.gov/table/ACSST1Y2021.S1810}{S1810} ~ With disability
\item \href{https://data.census.gov/table/ACSDT1Y2022.B11003?t=Family+Size+and+Type&g=040XX00US72&mode=results}{B11003} ~ Single parent
\item \href{https://data.census.gov/table/ACSDT1Y2022.B03002}{B03002} ~ Minority data
\item \href{https://data.census.gov/table/ACSDT1Y2021.C16001}{C16001} ~ English less than well
\item \href{https://data.census.gov/table/ACSDT5Y2020.B25024?q=B25024&g=500XX00US4829&table=B25024&tid=ACSDT5Y2020.B25024}{B25024} ~ Multi-unit housing(10+)
\item \href{https://data.census.gov/table/ACSDT5Y2020.B25024?q=B25024&g=500XX00US4829&table=B25024&tid=ACSDT5Y2020.B25024}{B25024} ~ Mobile housing
\item \href{https://data.census.gov/table/ACSDT5Y2021.B25014?q=B25014&g=1400000US48157672202}{B25014} ~ Crowding
\item \href{https://data.census.gov/table/ACSDT1Y2022.B25044}{B25044} ~ No vehicle
\item \href{https://data.census.gov/table/ACSDT1Y2023.B26001}{B26001} ~ Group quarters
}

4 themes and the 15 factors:
\itemize{
\item Factors of socioeconomic status: Below poverty, unemployment, income, no high school diploma(25+)
\item Factors of household composition and disability: 65 years old +, 17 years old -, civilians with disabilities, single-parent household
\item Factors of minority status and language: Minority, speaks English less than well
\item Factors of housing and transportation: Multi-unit structures, mobile homes, crowding, no vehicle, group quarters
}
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)
test_tib <- get_county_geo(test_tib)

add_SVI(test_tib)

}
\references{
Source:

Centers for Disease Control and Prevention (CDC). (2016, May 19). Social Vulnerability Index (SVI). \href{https://svi.cdc.gov/data-and-tools-download.html}{Retrieved from}.

General References:

Agency for Toxic Substances and Disease Registry. (2018, September). CDC’s Social Vulnerability Index (fact sheet). \href{https://svi.cdc.gov/factsheet.html}{Retrieved from}.
An, R., & Xiang, X. (2015). Social vulnerability and leisure-time physical inactivity among US adults. American Journal of Health Behavior, 39(6), 751–760.
Flanagan, B. E., Gregory, E. W., Hallisey, E. J., Heitgerd, J. L., & Lewis, B. (2011). A Social Vulnerability Index for disaster management. Journal of Homeland Security and Emergency Management, 8(1). doi: 10.2202/1547-7355.1792
Flanagan, B., & Hallisey, E. (2013). Social Vulnerability Index and Toolkit (PDF document). Presented at the New York State Department of Health and the Albany School of Public Health Annual GIS Day. \href{https://svi.cdc.gov/publications.html}{Retrieved from}.
Flanagan, B. E., Hallisey, E. J., Adams, E., Lavery, A. (2018). Measuring community vulnerability to natural and anthropogenic hazards: The Centers for Disease Control and Prevention’s Social Vulnerability Index. Journal of Environmental Health, 80(10), 34–36.
Gay, J. L., Robb, S. W., Benson, K. M., & White, A. (2016). Can the Social Vulnerability Index be used for more than emergency preparedness? An examination using youth physical fitness data. Journal of Physical Activity and Health, 13, 121–130.
Yee, C. W., Cunningham, S. D., & Ickovics, J. R. (2019). Application of the Social Vulnerability Index for identifying teen pregnancy intervention need in the United States. Maternal and Child Health Journal, 23, 1516–1524.
}
