% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{check_tidycensus_key}
\alias{check_tidycensus_key}
\title{Checks for Tidy Census Key}
\usage{
check_tidycensus_key()
}
\value{
TRUE if a Tidy Census Key was found and sends a message if otherwise
}
\description{
The function \code{check_tidycensus_key()} is used to check if the user has
obtained and entered a API key that would allow them to use the Tidy Census
package functions
}
\examples{
check_tidycensus_key()

}
