% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{clean_min_wage_data}
\alias{clean_min_wage_data}
\title{Cleans US Minimum Wage data}
\usage{
clean_min_wage_data(fed_min_wage = 7.25)
}
\arguments{
\item{fed_min_wage}{The current years minimum wage (e.g. current year: 2025, $7.25)}
}
\value{
A dataset that contains minimum wage data from 1968 to current year
with columns that contain individual years and states along with the
dichotomous and continuous measures of minimum wage.
}
\description{
The function \code{clean_min_wage_data()} takes excel data that is based on \href{https://www.dol.gov/agencies/whd/state/minimum-wage/history}{U.S. Department of Labor records} and
cleans it. The data is cleaned by filling in data for states that follow the
federal minimum wage, neglecting minimum wages that pertain to a very small
population, and deals with ranges of minimum wages. The method of deciding on
minimum wage was taking the over estimate because weighted average is not
possible without time stamps and in most states the higher minimum wage was
the majority of the population. This function also finds the dichotomous and
continuous measures for the minimum wage protocol.
}
\examples{
clean_min_wage_data(7.25)
DO NOT LEAVE PARAMETER EMPTY UNLESS YOU KNOW THE DEFAULT IS THE SAME AS THE CURRENT YEAR
*default = $7.25

}
