% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{county_state_helper}
\alias{county_state_helper}
\title{Gets County and State from NAME}
\usage{
county_state_helper(data_tib)
}
\arguments{
\item{data_tib}{A tibble of ACS data that needs \code{county} and \code{state} columns for future joins}
}
\value{
A tibble of given ACS data, but with added \code{county} and \code{state} columns
}
\description{
The \code{county_state_helper()} function is a helper function for getting the
\code{county} and \code{state} columns from the \code{NAME} column.
}
\examples{
per_capita <- tibble()
(Loops through all addresses)
{
  per_cap <- get_acs(
  geography = "tract",
  variables = "B19301_001",
  state = st,
  year = yr,
  survey = "acs5",
  output = "wide") \%>\% mutate(year = yr)
}
per_capita <- per_capita \%>\% bind_rows(per_cap)

county_state_helper(per_capita)

}
