% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geo_fallback}
\alias{geo_fallback}
\title{Backup geocoding method}
\usage{
geo_fallback(census_geo, fallback = "arcgis")
}
\arguments{
\item{census_geo}{A tibble that has been geocoded but still has NA values}

\item{fallback}{The method used to find coordinates of the missing values. By default this is "arcgis" because it showed to cover the most area compared to other methods}
}
\value{
A tibble with latitude and longitude values using arcgis to catch any missed values
}
\description{
The function \code{geo_fallback()} is responsible for finding the latitude and
longitude values of addresses that were not traceable by the census method
through the function \code{geocode()}.
}
\examples{
some_tibble <- some_tibble \%>\%
geocode(address, method = "census", lat = lat, long = long)

geo_fallback(some_tibble)

}
