% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{geo_info}
\alias{geo_info}
\title{Current Tibble Receives the Necessary Data}
\usage{
geo_info(tib)
}
\arguments{
\item{tib}{A tibble that contains at least \emph{address} column, plus any supplementary data}
}
\value{
A tibble with the address and corresponding data about that location(geographical divisions and geoids)
}
\description{
The \code{geo_info()} function is for when the input data is in tibble format
already. The function collects data such as geocoding & reverse geocoding
information and geoids for tract and county levels. The function finishes by
returning this newly updated tibble that is ready for the analysis in
following functions.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)

test_tib <- geo_info(test_tib)

}
