% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{get_county_geo}
\alias{get_county_geo}
\title{Collects county level GEOIDS for addresses}
\usage{
get_county_geo(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}
}
\value{
A tibble with prior data from the input tibble along with \code{GEOID_county}
}
\description{
The \code{get_county_geo()} function uses geocoding and spatial features in order to
obtain the GEOIDs for specific addresses. Using the \strong{Coordinate Reference
System(CRS) 4326}, \emph{cordinates} and \emph{state-year pairs} are created as \strong{sf
objects} and they are joined \strong{within} to get the results.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

test_tib <- keep_original_year(test_tib)
test_tib <- modify_year(test_tib, 2025)
test_tib <- get_GEOID(test_tib)
test_tib <- add_info_cols(test_tib)

get_county_geo(test_tib)

}
