% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{get_geodeterminants}
\alias{get_geodeterminants}
\title{Get the Social Determinants of Health for each Address}
\usage{
get_geodeterminants(
  gd_tib = NULL,
  gd_addresses = NULL,
  gd_current_year = 2025,
  gd_minority_group_code = NULL,
  gd_comparison_group_code = NULL,
  gd_minority_group_code_dhc = NULL,
  gd_minority_group_code_sf1 = NULL,
  gd_current_fed_min_wage = 7.25
)
}
\arguments{
\item{gd_tib}{\strong{Tibble} that must at least contain columns \code{address},
\code{state}(abbreviation for best use), and \code{year}(of diagnosis/source). This
tibble may also contain other patient data that will be preserved
throughout the function. Leave \strong{NULL} if data is in \emph{vector} form}

\item{gd_addresses}{\strong{Vector} named \code{addresses} that contains addresses of
patient/study subjects}

\item{gd_current_year}{The current year this function is being used}

\item{gd_minority_group_code}{The \code{B03002} table code of the minority group
of interest. If the code isn't known leave null and select from the menu.
\href{https://data.census.gov/table/ACSDT1Y2022.B03002}{source}}

\item{gd_comparison_group_code}{The \code{B03002} table code of the comparison
group of interest. If the code isn't known leave null and select from the
menu. \href{https://data.census.gov/table/ACSDT1Y2022.B03002}{source}}

\item{gd_minority_group_code_dhc}{The \code{P5} table code of the comparison group
of interest. If the code isn't known leave null and select from the menu.
Used when dates are 2020 and up. \href{https://www.census.gov/data/tables/2023/dec/2020-census-dhc.html}{DHC data}}

\item{gd_minority_group_code_sf1}{The \code{P005} table code of the comparison
group of interest. If the code isn't known leave null and select from the
menu. Used when dates are below 2020. \href{https://www.census.gov/data/datasets/2010/dec/summary-file-1.html}{sf1 data}}

\item{gd_current_fed_min_wage}{The current years minimum wage (e.g. current
year: 2025, $7.25)}
}
\value{
The \strong{original} input tibble with \strong{Social Determinants of
Health(SDOH)} data appended for each address if data is available
according to each determinants' function criteria
}
\description{
Glossary for variables and functions: \href{https://docs.google.com/document/d/1r7E8aK60fNnDKe8ZtFaBNqZ3VAQXjU62YCb9vlPllfk/edit?usp=sharing}{Glossary}
⚠️ \strong{Note:} Data must be in tibble or vector form, read parameter descriptions for formatting break down.
You must have these packages installed for use of this package/function.
\code{\link{Street}}, \code{\link{City}}, \code{\link{State}} \code{\link{Area Code}} format works best for address.
Download and unzip the file from github for access to the required data \href{https://wchan05.github.io/data_access/}{here}.

Converts a vector or tibble of addresses and supplementary data into Census
tract and block identifiers using \pkg{tidygeocoder}, and appends
tract-level measures of social determinants of health (SDOH) from
authoritative public data sources (e.g. ACS and EPA). Designed for public
health and policy research, this function provides a interlinking workflow for
joining address-level data to environmental, socioeconomic, and
demographic indicators.

These packages must be installed for proper functionality:
\strong{tidyverse}, \strong{tidygeocoder}, \strong{tibble}, \strong{tidycensus}(API KEY), \strong{tigris}, \strong{sf}, \strong{readxl}

SDOH list that can be retrieved given strictly address data:
Air Quality Index, Concentrated Poverty, Educational Attainment – Community,
Environmental Justice, Food Swamp, Income Concentration, Minimum Wage,
Percent Unionized for Non-Agricultural Labor Force, Race/Ethnic Residential
Segregation – American Community Survey, Race/Ethnic Residential Segregation
– Separation (S) Index, Unbiased, Race/Ethnic Residential Segregation – U.S.
Census, Social Vulnerability
}
\examples{
test_tib <- tibble(
address = c(
  "15 Main Street Flemington, NJ 08822",
  "120 E Main St, Ramsey, MN 55303",
  "1315 S Glenstone Ave, Springfield, MO 65804",
  "401 W 14th St, Austin, TX 78701",
  "340 S Lemon Ave, Walnut, CA 91789",
  "25 Main St, Northampton, MA 01060",
  "Non existent Ave",
  "100 Jersey Ave, New Brunswick, NJ 08901",
  "17350 Woodward Ave, Detroit, MI 48203",
  "17350 Woodward Ave, Detroit, MI 48203",
  "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

view(geodeterminants(test_tib))

addresses <- c(
  "1600 Pennsylvania Ave NW, Washington, DC 20500",
  "1 Infinite Loop, Cupertino, CA 95014",
  "350 5th Ave, New York, NY 10118",
  "233 S Wacker Dr, Chicago, IL 60606",
  "15 Main Street Flemington, NJ 08822",
  "Non existent Ave")

view(geodeterminants(gd_addresses = addresses))
}
\references{
"Structural Social Determinants of Health." PhenX Toolkit, Version 48.2, RTI International, 17 July 2025, \url{https://www.phenxtoolkit.org/sub-collections/view/32}.
}
