% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{keep_original_year}
\alias{keep_original_year}
\title{Preserves year of diagnosis from the input data}
\usage{
keep_original_year(tib)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data}
}
\value{
A tibble of the original data, except the \emph{year} column(which
represents the year of diagnosis) is copied and stored as the column
\emph{actual_year}
}
\description{
The \code{keep_original_year()} function will save the \strong{original diagnosis year}
from the input data in a separate column, so that when the years are modified
to assure available and accurate data, the actual year is \strong{NOT OVERWRITTEN}.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

keep_original_year(test_tib)

}
