% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{look_up_helper}
\alias{look_up_helper}
\title{Gets state abbreviations}
\usage{
look_up_helper(data_tib)
}
\arguments{
\item{data_tib}{A tibble of ACS data that has a state column}
}
\value{
A tibble of ACS data with the state column changed to \strong{state abbreviations} \emph{instead} of \strong{full names}
}
\description{
The function \code{look_up_helper()} gets the \strong{abbreviation} of state names as it is
more compatible for joining the data found
}
\examples{
look_up_helper(any tibble with a state column)

}
