% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{modify_year}
\alias{modify_year}
\title{Modifies years to provide attainable and accurate data}
\usage{
modify_year(tib, current_year)
}
\arguments{
\item{tib}{A tibble of data containing addresses as \code{address} in characters
and any corresponding data. The tibble should at least include addresses in
proper format and year of diagnosis as \code{year}(until function
\code{keep_original_year()} is run which then \code{actual_year} and \code{year} exist)
for the most accurate data.}

\item{current_year}{A 4 digit value that can represent the current year data
is being used with this package or the year data was finalized}
}
\value{
A tibble with the \emph{address}, \emph{actual_year}, and \emph{year} columns as
well as any other columns that were in the input tibble.
}
\description{
⚠️ \strong{Note:} It is important to run \code{keep_original_year()}
before this function so that year data is not overwritten.

The \code{modify_year} function takes the year of diagnosis column and adjusts the
years to assure that attainable and accurate data is collect. The reason for
this is because not every year is supported by ACS dataset and other files.
2012 is the earliest data we are using because that is when full datasets
were released and the beginning of the more accurate ones. ACS data tends to
have full data up to 2 years before the current year(Ex. current year is 2025
so 2023 will be the most recent full data). Using this method that is how we
set the default year when year of diagnosis is NA or invalid.
}
\examples{
test_tib <- tibble(
address = c(
 "15 Main Street Flemington, NJ 08822",
 "120 E Main St, Ramsey, MN 55303",
 "1315 S Glenstone Ave, Springfield, MO 65804",
 "401 W 14th St, Austin, TX 78701",
 "340 S Lemon Ave, Walnut, CA 91789",
 "25 Main St, Northampton, MA 01060",
 "Non existent Ave",
 "100 Jersey Ave, New Brunswick, NJ 08901",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203",
 "17350 Woodward Ave, Detroit, MI 48203"),
state = c("NJ", "MN", "MO", "TX", "CA", "MA", "PA", "NJ", "MI", "MI", "MI"),
year = c(2022, 2023, NA, 2023, NA, 2023, 2022, 2022, NA, 1990, 2020))

modify_year(keep_original_year(test_tib))

}
