% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{SDOH_reverse}
\alias{SDOH_reverse}
\title{Appends geographical categorical information}
\usage{
SDOH_reverse(geo_tib)
}
\arguments{
\item{geo_tib}{A tibble containing addresses and geocoding information,
returned from \code{vector_geocode_to_tib()}}
}
\value{
A tibble of addresses, their geocoded information, and general
information about their location including: \code{city}, \code{state}, \code{county}, and
\code{country}
}
\description{
The \code{SDOH_reverse()} function is used to fill in missing information about an
\emph{address} and for joining data later on. The intent is to be used after
\code{vector_geocode_to_tib()}. Information like \emph{city}, \emph{state}, \emph{county}, and
\emph{country} for a location will be appended.
}
\examples{
addresses <- c(
"1600 Pennsylvania Ave NW, Washington, DC 20500",
"1 Infinite Loop, Cupertino, CA 95014",
"350 5th Ave, New York, NY 10118",
"233 S Wacker Dr, Chicago, IL 60606",
"15 Main Street Flemington, NJ 08822",
"Non existent Ave")

geocoded <- vector_geocode_to_tib(addresses)

SDOH_reverse(geocoded)
}
