% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{vector_geocode_to_tib}
\alias{vector_geocode_to_tib}
\title{Convert vector of addresses to compatible tibble}
\usage{
vector_geocode_to_tib(addresses)
}
\arguments{
\item{addresses}{A vector of addresses}
}
\value{
A tibble that has the addresses from the original vector and
geocoded information for those that it is available.
}
\description{
The use of this function is for when only a vector of addresses is provided.
The \code{vector_geocode_to_tib()} function first validates that the vector is in
proper format. It checks if the vector is empty, if all string values are
present, and warns the user about NA values. The addresses are then put into
a tibble, geocoded, and returned in tibble data.
}
\examples{
addresses <- c(
"1600 Pennsylvania Ave NW, Washington, DC 20500",
"1 Infinite Loop, Cupertino, CA 95014",
"350 5th Ave, New York, NY 10118",
"233 S Wacker Dr, Chicago, IL 60606",
"15 Main Street Flemington, NJ 08822",
"Non existent Ave")

vector_geocode_to_tib(addresses)
}
