% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geocoding.R
\name{vector_to_tib}
\alias{vector_to_tib}
\title{Vector to Tibble with Information}
\usage{
vector_to_tib(addresses, current_year)
}
\arguments{
\item{addresses}{A vector of addresses}

\item{current_year}{The current year}
}
\value{
A tibble with the address and corresponding data about that location(geographical divisions and geoids)
}
\description{
The \code{vector_to_tib()} function is a function that takes the address vector
and makes it into a tibble format, collecting information such as setting
default dates, geocoding & reverse geocoding information, and geoids for
tract and county levels. The function finishes by tying all the data together
into a tibble that originated from the vector of addresses and returns it.
The purpose of transferring to a tibble is because it is required for
obtaining the social determinants of health and other functions.
}
\examples{
#' addresses <- c(
"1600 Pennsylvania Ave NW, Washington, DC 20500",
"1 Infinite Loop, Cupertino, CA 95014",
"350 5th Ave, New York, NY 10118",
"233 S Wacker Dr, Chicago, IL 60606",
"15 Main Street Flemington, NJ 08822",
"Non existent Ave")

geocoded <- vector_to_tib(addresses)

}
